LuxSimpleChainsExt: specify rng when initializing
Propagates the RNG used by Lux.setup() to SimpleChains via the adaptor.
This makes the SimpleChains-backed network setup behavior match that of
standard Lux networks.

While this keyword argument is not currently documented in SimpleChains
the author has stated in a response to a question whether it should be
considered public API that it is.

Closes #554.